requirements.txt
dateparser
datetime
requests.adapters
json
pandas
numpy
BeautifulSoup
fuzzywuzzy
elasticsearch
TextBlob
